backend/requirements.txt
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
better-profanity==0.7.0
CacheControl==0.14.4
cbor2==5.9.0
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.6
constantly==23.10.4
cryptography==46.0.6
daphne==4.2.1
Django==6.0.3
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.17.1
firebase_admin==7.3.0
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-core==2.5.0
google-cloud-firestore==2.26.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.1
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
iniconfig==2.3.0
msgpack==1.1.2
packaging==26.0
pluggy==1.6.0
proto-plus==1.27.2
protobuf==6.33.6
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.20.0
PyJWT==2.12.1
pyOpenSSL==26.0.0
pytest==9.0.3
python-dotenv==1.2.2
redis==7.4.0
requests==2.33.0
service-identity==24.2.0
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
ujson==5.12.0
urllib3==2.6.3
zope.interface==8.3
